+++ src/arch/x86_64/kernel/mmio.rs
@@ -11,7 +11,7 @@ use crate::collections::irqsave;
 use crate::arch::x86_64::mm::paging;
 use crate::arch::x86_64::mm::paging::{BasePageSize, PageSize, PageTableEntryFlags};
 use crate::arch::x86_64::mm::PhysAddr;
-use crate::drivers::net::virtio_mmio::VirtioNetDriver;
+use crate::drivers::net::virtio_net::VirtioNetDriver;
 use crate::drivers::net::NetworkInterface;
 use crate::drivers::virtio::transport::mmio as mmio_virtio;
 use crate::drivers::virtio::transport::mmio::{DevId, MmioRegisterLayout, VirtioDriver};
